WIP: hi-res 640x400 renderer (Stage 1) + limit removal
Bundles the (previously uncommitted) renderer + P limit removal with
Stage 1 of the hi-res renderer. Local WIP checkpoint on branch wip/hires;
NOT pushed.

Hi-res Stage 1 (DONE, verified via screenshot on DOOM2 MAP01):
- i_video.h: ORIGWIDTH/ORIGHEIGHT (320x200) split from SCREENWIDTH/HEIGHT
  (640x400) via HIRES factor; native 640x400 render buffer, 1:1 blit.
- r_defs.h: visplane top/bottom/pads byte -> unsigned short (byte overflows
  past y=255 at 400 tall).
- r_plane.c: visplane sentinel 0xff -> 0xffff.
- r_main.c / r_draw.c: view-size and SBARHEIGHT math scaled by HIRES.
- i_video.c: -shotquit N screenshot harness (dumps framebuffer to BMP).

REMAINING (2D layer): Stage 3 scale V_DrawPatch & siblings + hi-res the
status bar (st_backing_screen/V_CopyRect); Stage 4 fix screen wipe
(f_wipe.c); audit UI centering SCREENWIDTH->ORIGWIDTH.

TEMP HACKS to replace in Stage 3: st_lib.c three range-checks changed from
I_Error to early-return so the (not-yet-scaled) status bar doesn't crash.

+++ doomgeneric/r_defs.h
@@ -429,16 +429,15 @@ typedef struct
   int			maxx;
   
   // leave pads for [minx-1]/[maxx+1]
-  
-  byte		pad1;
-  // Here lies the rub for all
-  //  dynamic resize/change of resolution.
-  byte		top[SCREENWIDTH];
-  byte		pad2;
-  byte		pad3;
-  // See above.
-  byte		bottom[SCREENWIDTH];
-  byte		pad4;
+  // Widened from byte to unsigned short for hi-res: at SCREENHEIGHT=400 a
+  // y-coordinate (and the 0xffff "no span" sentinel) no longer fits in a byte.
+
+  unsigned short	pad1;
+  unsigned short	top[SCREENWIDTH];
+  unsigned short	pad2;
+  unsigned short	pad3;
+  unsigned short	bottom[SCREENWIDTH];
+  unsigned short	pad4;
 
 } visplane_t;
